+++ tephi_module/_constants.py
